+++ src/module.cpp
@@ -604,7 +604,7 @@ namespace module
 			const std::string contents = sys::read_file(file);
 
 			attr[variant("md5")] = variant(md5::sum(contents));
-			attr[variant("size")] = variant(contents.size());
+			attr[variant("size")] = variant(static_cast<int>(contents.size()));
 
 			auto attr_copy = attr;
 
@@ -628,7 +628,7 @@ namespace module
 			const std::string contents = variant(&manifest_file).write_json();
 
 			attr[variant("md5")] = variant(md5::sum(contents));
-			attr[variant("size")] = variant(contents.size());
+			attr[variant("size")] = variant(static_cast<int>(contents.size()));
 
 			std::vector<char> data(contents.begin(), contents.end());
